+++ test/libsolidity/syntaxTests/comments/multiline_doc_unicode_direction_override_2.sol
@@ -1,0 +1,9 @@
+contract C {
+    function f() public pure
+    {
+        // PDF
+        /** underflow ‬ */
+    }
+}
+// ----
+// ParserError 8936: (71-85): Unicode direction override underflow in comment or string literal.