+++ test/libsolidity/syntaxTests/comments/singleline_doc_unicode_direction_override_1.sol
@@ -1,0 +1,9 @@
+contract C {
+    function f() public pure
+    {
+        // RLO
+        /// overflow ‮
+    }
+}
+// ----
+// ParserError 8936: (71-92): Mismatching directional override markers in comment or string literal.