+++ test/libsolidity/syntaxTests/comments/singleline_doc_unicode_direction_override_3.sol
@@ -1,0 +1,8 @@
+contract C {
+    function f() public pure
+    {
+        // RLO PDF
+        /// ok ‮‬
+    }
+}
+// ----